Pkg restructure: util_computer.py -> computer/
